chore: move vault to pkg
