+++ internal/migration/up/up.go
@@ -11,6 +11,7 @@ import (
 	"github.com/spf13/afero"
 	"github.com/supabase/cli/internal/utils"
 	"github.com/supabase/cli/pkg/migration"
+	"github.com/supabase/cli/pkg/vault"
 )
 
 func Run(ctx context.Context, includeAll bool, config pgconn.Config, fsys afero.Fs, options ...func(*pgx.ConnConfig)) error {
@@ -21,6 +22,9 @@ func Run(ctx context.Context, includeAll bool, config pgconn.Config, fsys afero.
 	defer conn.Close(context.Background())
 	pending, err := GetPendingMigrations(ctx, includeAll, conn, fsys)
 	if err != nil {
+		return err
+	}
+	if err := vault.UpsertVaultSecrets(ctx, utils.Config.Db.Vault, conn); err != nil {
 		return err
 	}
 	return migration.ApplyMigrations(ctx, pending, conn, afero.NewIOFS(fsys))
